fix: checkpoint diff type need be float64

+++ app/service/milestone.entity.go
@@ -24,7 +24,7 @@ type CheckpointCreate struct {
 
 type CheckpointView struct {
 	model.Checkpoint
-	Diff int
+	Diff float64
 }
 
 type CheckpointQuery struct {
